+++ yaml/src/test/java/com/mrivanplays/annotationconfig/yaml/LocationNoConfigObject.java
@@ -1,0 +1,30 @@
+package com.mrivanplays.annotationconfig.yaml;
+
+public class LocationNoConfigObject {
+
+  private String world;
+  private int x, y, z;
+
+  public LocationNoConfigObject(String world, int x, int y, int z) {
+    this.world = world;
+    this.x = x;
+    this.y = y;
+    this.z = z;
+  }
+
+  public String getWorld() {
+    return world;
+  }
+
+  public int getX() {
+    return x;
+  }
+
+  public int getY() {
+    return y;
+  }
+
+  public int getZ() {
+    return z;
+  }
+}